Checked for breaking code, seems fine
added some messages for registering that popup if passwords do not match, or invalid email. had to change form type to text to actually use regex to check and alert user. it seems that it being an email type input the browser checks if it is an email and stops anything from happening, which also prevents messages from appearing